+++ src/services/Customers.php
@@ -14,7 +14,6 @@ use craft\commerce\elements\Order;
 use craft\commerce\models\Address;
 use craft\commerce\models\Customer;
 use craft\commerce\Plugin;
-use craft\commerce\queue\jobs\ConsolidateGuestOrders;
 use craft\commerce\records\Customer as CustomerRecord;
 use craft\commerce\records\CustomerAddress as CustomerAddressRecord;
 use craft\commerce\web\assets\commercecp\CommerceCpAsset;
@@ -427,9 +426,7 @@ class Customers extends Component
         }
 
         // Consolidate guest orders
-        Craft::$app->getQueue()->push(new ConsolidateGuestOrders([
-            'emails' => [$order->email]
-        ]));
+        $this->consolidateGuestOrdersByEmail($order->email);
     }
 
     /**
@@ -544,6 +541,53 @@ class Customers extends Component
         }
 
         return $customersQuery;
+    }
+
+    /**
+     * Consolidate all guest orders for this email address to use one customer record.
+     *
+     * @param string $email
+     * @throws \yii\db\Exception
+     * @since 3.x
+     */
+    public function consolidateGuestOrdersByEmail(string $email)
+    {
+        $customerId = (new Query())
+            ->select('orders.customerId')
+            ->from(Table::ORDERS . ' orders')
+            ->innerJoin(Table::CUSTOMERS . ' customers', '[[customers.id]] = [[orders.customerId]]')
+            ->where(['orders.email' => $email])
+            ->andWhere(['orders.isCompleted' => true])
+            // we want the customers related to a userId to be listed first, then by their latest order
+            ->orderBy('[[customers.userId]] DESC, [[orders.dateOrdered]] ASC')
+            ->scalar(); // get the first customerId in the result
+
+        if (!$customerId) {
+            return;
+        }
+
+        // Get completed orders for other customers with the same email but not the same customer
+        $orders = (new Query())
+            ->select([
+                'id' => 'orders.id',
+                'userId' => 'customers.userId'
+            ])
+            ->where(['and', ['[[orders.email]]' => $email, '[[orders.isCompleted]]' => true], ['not', ['[[orders.customerId]]' => $customerId]]])
+            ->leftJoin(Table::CUSTOMERS . ' customers', '[[orders.customerId]] = [[customers.id]]')
+            ->from(Table::ORDERS . ' orders')
+            ->all();
+
+        foreach ($orders as $order) {
+            $userId = $order['userId'];
+            $orderId = $order['id'];
+
+            if (!$userId) {
+                // Dont use element save, just update DB directly
+                Craft::$app->getDb()->createCommand()
+                    ->update(Table::ORDERS . ' orders', ['[[orders.customerId]]' => $customerId], ['[[orders.id]]' => $orderId])
+                    ->execute();
+            }
+        }
     }
 
     /**